+++ lib/domino.dart
@@ -15,11 +15,25 @@ abstract class BuildContext {
 /// Builds a single or List-embedded structure of Nodes and/or Components.
 typedef dynamic BuildFn(BuildContext context);
 
-/// A state-holder component that builds a single or List-embedded structure of
-/// Nodes and/or Components.
+/// Builds a single or List-embedded structure of Nodes and/or Components.
 abstract class Component {
   /// Builds a single or List-embedded structure of Nodes and/or Components.
   dynamic build(BuildContext context);
+}
+
+/// Stateful (but disposable) [Component] that provides lifecycle hooks to store
+/// and retrieve its state.
+abstract class StatefulComponent<S> extends Component {
+  /// Gets the state of the component.
+  ///
+  /// The getter will be called only before the a rebuild requires it.
+  S getState();
+
+  /// Set the state of the component.
+  ///
+  /// The setter will be called right before [build] is called. [value] will be
+  /// either the value from the previous instance's [getState] call, or null.
+  void setState(S value);
 }
 
 /// Provides lifecycle handling for a hierarchy of components.